% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cms_data}
\alias{cms_data}
\title{Charlotte-Mecklenberg School district in the 2018-2019 school year.}
\format{
A dataframe with 170 rows and 7 columns.
}
\usage{
cms_data
}
\description{
A dataset containing school-level information for CMS schools, including demographic breakdowns
   by school and school grades.
}
\keyword{datasets}
