% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ipg_data}
\alias{ipg_data}
\title{IPG scores by classroom.}
\format{
A dataframe with 100 rows and 14 columns.
}
\usage{
ipg_data
}
\description{
A dataset containing IPG scores by classroom. The data is simulated and not actual data.
}
\keyword{datasets}
