% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_metric.R
\name{make_metric}
\alias{make_metric}
\title{Calculate raw common metric values}
\usage{
make_metric(data, metric, scaleusewarning = T)
}
\arguments{
\item{data}{Data from a single timepoint. Used in \code{tntpmetric_mean}.}

\item{metric}{Quoted name of the common metric. Options are "engagement", "belonging",
"relevance", "assignments", "tntpcore", or "ipg".}

\item{scaleusewarning}{A logical (T/F) indicating whether function should generate a warning when
not all values of a scale are used. For example, student survey data that only contains values
of 1s and 2s could mean that data is on a 1-4 scale, when it should be on a 0-3 scale. When
\code{scaleusewarning = T}, the function will warn you of this. This warning does not mean your
data is wrong. For example, the Academic Ownership domain from TNTP CORE has 5 potential
values: 1, 2, 3, 4, or 5. It's not uncommon to have data where teachers were never rated above
a 4 on this domain. In this case, the printed warning can be ignored. Default if \code{TRUE}.
If you are confident your data is on the right scale, you can suppress the warning by setting
to \code{TRUE}.}
}
\value{
A data.frame identical to the original except with a new column/variable named as the
  input to metric with a cm_ prefic (e.g., \code{cm_engagement}, \code{cm_ipg}, etc.) that has
  the value of the scored metric. Currently, attirbutes of original data.frame (like groups) are
  not preserved.
}
\description{
\code{make_metric} takes raw data containing all of the indicators, survey items, etc. needed
  to calculate a score on a common metric. It creates a new column/variable in your data with the
  value of the scored metric.
}
\section{Data and Variable Format}{

  \code{make_metric} should be used with the raw metric data.
  Each row of data should represent a single rated outcome. For example, each row of data will be
  a single completed survey, a single rated assignment, a single classroom observation, etc.
  The data must have the components needed to score the construct. For example, data on student
  engagement should have variables corresponding to the four survey questions used to calculate
  engagement. Leave all items in their raw form - the functions automatically account for items
  that need to be reverse coded (if any). The only requirement is that the data contains the
  needed variables and that the variables are numeric (i.e., data values should be 0s and 1s,
  not 'No' and 'Yes'. This ensures that the common metrics are calculated correctly and
  consistently across projects. Each metric has its own set of needed variables that must be
  spelled exactly as shown below. They are:
  \describe{
    \item{engagement:}{eng_like, eng_losttrack, eng_interest, eng_moreabout}
    \item{belonging:}{tch_problem, bel_ideas, bel_fitin, tch_interestedideas}
    \item{relevance:}{rel_asmuch, rel_future, rel_outside, rel_rightnow}
    \item{expectations:}{exp_mastergl, exp_toochallenging, exp_oneyear, exp_different,
      exp_overburden, exp_began}
    \item{tntpcore:}{ec, ao, dl, cl}
    \item{ipg:}{ca1_a, ca1_b, ca1_c, ca2_overall, ca3_overall, col, rfs_overall}
   }
 See the vignette for more details.
}

\examples{
# Compute the engagement score for each collected survey
make_metric(ss_data_initial, metric = "engagement")

}
