% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ss_data_final}
\alias{ss_data_final}
\title{Final student survey.}
\format{
A dataframe with 1000 rows and 16 columns.
}
\usage{
ss_data_final
}
\description{
A dataset containing simulated data from a final student survey. It contains the same columns as the ss_survey_initial dataset.
}
\keyword{datasets}
