% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ss_data_initial}
\alias{ss_data_initial}
\title{Initial student survey.}
\format{
A dataframe with 1000 rows and 16 columns.
}
\usage{
ss_data_initial
}
\description{
A dataset containing simulated data from an initial student survey. It contains the same columns as the ss_survey_final dataset.
}
\keyword{datasets}
